Switch currencies from AUD to USD
I hadn't initially realised I could charge people in USD whilst linking
to an Australia bank account. I'd prefer to bill in USD as most of the
running costs for GopherCI is in USD anyway.

There's an edge case though, where existing customers need to be
allocated a new Stripe customer ID, as Stripe won't allow an existing
customer to have subscriptions (even though one is disabled) in
different currencies.

So, for those customers, we'll just create a new Stripe customer for
them when they change plans, and eventually, when everyone is off of
the old PersonalMonthly plan (and onto PersonalMonthlyUSD or similar)
we can just delete those plans. The only side effect is those customers
won't be able to see their existing, cancelled, subscriptions, but
we won't delete the customer in Stripe.