Implement checks for detecting possible double charging.
This new command, `gopherci billing:check` will need to be
executed via some kind of cron and outputs to stdout.

Also included is a minor framework for additional commands to be
added in the future. It's very brief and will likely need a
refactor.

Migrations continue to be automatic, just like in gopherci, it's
not needed for this project, but I've kept that behaviour consistent
as there's little cost to doing it, and makes gopherci easier to
upgrade for others. Rollback was refactored into the new commands
package.

I haven't included tests simply due to time.

Resolves #10.