Replace user scope with user:email for GitHub OAuth
Previously we've been asking for access to the user scope, but I
don't believe I need it at all after some further testing, it's
just the user's email.

I will still need read only access to organisations, in order to
list all organisations so the user can enable an installation. But
there may be a workflow that could enable us to remove this scope
too.

Further, this change improves error handling when no tokens exist
for a user, previously we assumed a token always existed (even though
the database field is null, as we may support GitLab OAuth later),
and we passed that value to json.Unmarshal, which returned an error
and this resulted in a 500 Internet Server Error for the user with
no recourse.

Now, we only unmarshal the token if it's not nil (not null). If it
is nil, we provide the OAuth client an empty token (no bearer and no
refresh) which, when invoked by the google/go-github client, causes
it to return an error to the client and eventually our invocation. It
now becomes our responsibility to handle this error, and we'll choose
to attempt an oauth flow again (there is a chance this could end in
an infinite loop, but this hasn't been addressed).

Finally, because we don't need the previous permissions anymore, an
additional migration is provided to remove all tokens. On next login
the user can add the new permissions.